2013-01-28. Pushing a current log. HumMod version bumped to 2.0.41.
HumMod previously used a linear predictor to predict fatty acid vs. glucose use in each of the major organs. This worked OK with normal deliveries but not with unusual
deliveries -- specifically brain ketoacid use during startvation.

So I replaced this predictor with a non-linear one that is much better. The expense is that fatty acid preference is no longer a working parameter. There is a detailed reference in Structure\FuelSelector. The new predictor is visible at 'Organ' | Fuel for each organ.