2013-02-14 HumMod v2.0.48. Pushing current version of log. Starting to implement John Clemmer's work and detailed amino acid metabolism.
Expanded autopsy report to include both examination and chemistry.

In posture control and energy and status, a constant was named both TILTED and TILTING, eliciting an unknown variable error when HumMod dies. Fixed to TILTING for all references.

Put lots of note files on amino acids in Structure\AminoAcid\Notes. These are references for future expansion of amino acid physiology.